Create Code to Write name using Command Line Argument
Since we are using the Command Line Argument method, we do not need to convert from String to any other version when dealing with codes related to string, and hence we directly use f1=(args[0]) and not like  any other format that we may use for Integer, double, float etc. i.e n1=Integer.parseInt(args[0]).